Make consent component event
